pyproject.toml
[tool.poetry]
name = "crypto"
version = "0.1.0"
description = ""
authors = ["Palon PL Mung <Palon.PL.Mung@ibm.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
allure-python-commons = "^2.13.5"
appium-python-client = "2.11.1"
jsonpath = "^0.82.2"
logger = "^1.4"
pytest = "^8.3.4"
pytest-bdd = "^8.1.0"
requests = "^2.32.3"
selenium = "^4.29.0"
pytest-html = "^4.1.1"
pyyaml = "^6.0.2"
pytest-xdist = "^3.6.1"
allure-pytest = "^2.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
